requirements.txt
langgraph
langchain-core
langchain-openai
psycopg2-binary
